Start process of converting to a fully-parameterised product.
Added issues checkboxes to the form in a new Fieldset, and a
(so far untested) issues TextField to the Feedback record.
Need to investigate the Django way to store this stuff.
Added defaults to the form's creation parameters. Note that
the form class is passed in as a form class attribute, so
there is no chance to modify the object creation code. We
should therefore use closures or some such trickery to bake
the parameter values into the form class.
I also shortened the rather cumbersome getattr method of
the DjangoFeedbackGovUkSettings class somewhat.